refactor: simplify project structure
