+++ README.md
@@ -52,7 +52,7 @@ For more clarity on this project and its structure you can also check out the de
 ## License
 
 ```
-   Copyright 2013-2023 Prismic <contact@prismic.io> (https://prismic.io)
+   Copyright 2013-2024 Prismic <contact@prismic.io> (https://prismic.io)
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
