+++ src/PrismicRichText.ts
@@ -81,8 +81,7 @@ export type PrismicRichTextProps = {
 	htmlSerializer?: HTMLFunctionSerializer | HTMLMapSerializer
 
 	/**
-	 * An HTML tag name, a component, or a functional component used to wrap the
-	 * output.
+	 * An HTML tag name or a component used to wrap the output.
 	 *
 	 * @defaultValue `"div"`
 	 */
