+++ src/types.ts
@@ -46,8 +46,7 @@ type PrismicPluginComponentsOptions = {
 	linkRel?: AsLinkAttrsConfig["rel"]
 
 	/**
-	 * An HTML tag name, a component, or a functional component used to render
-	 * internal links.
+	 * An HTML tag name or a component used to render internal links.
 	 *
 	 * @remarks
 	 * HTML tag names will be rendered using the anchor tag interface (`href`,
@@ -61,8 +60,7 @@ type PrismicPluginComponentsOptions = {
 	linkInternalComponent?: ComponentOrTagName
 
 	/**
-	 * An HTML tag name, a component, or a functional component used to render
-	 * external links.
+	 * An HTML tag name or a component used to render external links.
 	 *
 	 * @remarks
 	 * HTML tag names will be rendered using the anchor tag interface (`href`,
